requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
mysql-connector-python==8.0.26
google-generativeai==0.3.0
pydantic==1.8.2 